requirements.txt
matplotlib==3.4.3
matplotlib-inline==0.1.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
plotly==5.3.1
streamlit==1.0.0